+++ apps/chaingraph-frontend/src/store/trpc/store.ts
@@ -17,7 +17,10 @@ export const createTRPCClientEvent = trpcDomain.createEvent<{
 }>()
 
 export const $trpcClient = trpcDomain.createStore<TRPCClient | null>(null)
-  .on(createTRPCClientEvent, (_, { trpcURL, superjsonCustom }) => {
+  .on(createTRPCClientEvent, (s, { trpcURL, superjsonCustom }) => {
+    if (s !== null) {
+      return s
+    }
     return createTRPCClient({
       url: trpcURL ?? `ws://localhost:3001`,
       superjsonCustom: superjsonCustom ?? SuperJSON,