+++ apps/chaingraph-backend/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @badaitech/chaingraph-backend
 
+## 0.1.6-dev.43
+
+### Patch Changes
+
+- Fix ui logic and effector store execution reset trigger
+- Updated dependencies
+  - @badaitech/chaingraph-nodes@0.1.6-dev.43
+  - @badaitech/chaingraph-types@0.1.6-dev.43
+  - @badaitech/chaingraph-trpc@0.1.6-dev.43
+
 ## 0.1.6-dev.42
 
 ### Patch Changes